Use Arranger Router rc4, use separate pm2 file, update comments
