+++ cms/package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@aws-sdk/credential-provider-node": "^3.972.66",
     "@opensearch-project/opensearch": "^3.6.0",
-    "@overture-stack/arranger-graphql-router": "1.0.0-rc.3",
+    "@overture-stack/arranger-graphql-router": "1.0.0-rc.4",
     "@overture-stack/arranger-types": "1.0.0-rc.2",
     "aws-sdk": "^2.723.0",
     "axios": "^1.16.0",
